+++ src/components/Button.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Button = ({ innerText, clicked }) => {
+   return <button onClick={clicked}>{innerText}</button>
+};
+
+export default Button;